added ecommerce impressions, fixed analytics keyword, refactoring

+++ src/Managed/GoogleAnalytics.Core/Ecommerce/Product.cs
@@ -16,7 +16,7 @@ namespace GoogleAnalytics.Ecommerce
         public Product()
         {
             CustomDimensions = new Dictionary<int, string>();
-            CustomMetrics = new Dictionary<int, int>();
+            CustomMetrics = new Dictionary<int, float>();
         }
 
         /// <summary>
@@ -42,7 +42,7 @@ namespace GoogleAnalytics.Ecommerce
         /// <summary>
         ///     Gets or sets the custom metrics associated with the product.
         /// </summary>
-        public IDictionary<int, int> CustomMetrics { get; }
+        public IDictionary<int, float> CustomMetrics { get; }
 
         /// <summary>
         ///     Gets or sets the id that is used to identify a product in GA reports.
